style : 반응형 설계 완료
로딩시에 스켈레톤 나오도록 추가기능 필요
마우스핸들링과 웹페이지 width받아오는 커스텀훅 분리
