+++ src/pages/list/CardListPage.scss
@@ -3,83 +3,94 @@
 @import '../../assets/styles/palette.scss';
 
 .card-list-layer {
-  margin: 0 auto;
+  margin: 50px auto;
   width: 100%;
   overflow: hidden;
   @include Flex(column, center, center, null);
 
-  // @include tablet {
-  //   .card-list-box {
-  //     width: 100%;
-  //     .card-list-data {
-  //       width: 100%;
-  //       .arrow {
-  //         display: none;
-  //       }
-  //       .card-list-wrapper {
-  //       }
-  //     }
-  //   }
-  // }
-
-  // @include mobile {
-  // }
+  @include mobile {
+    margin-top: 40px
+  }
 
   .card-list-box {
-    @include Flex(column, center, flex-start, null);
-    // padding: 0 24px;
+    @include Flex(column, center, center, null);
     width: 100%;
     margin-bottom: 50px;
 
-    .card-list-title {
-      margin: 0 24px;
-      margin-bottom: 16px;
-      @include font-style(24, bold);
-      line-height: 36px;
+    @include mobile {
+      margin-bottom: 74px;
+    }
+
+    .title-layer {
+      @include Flex(column, flex-start, flex-start, null);
+      width: 100%;
+      max-width: 1200px;
+      .card-list-title {
+        margin: 0 24px 16px;
+
+        @include font-style(24, bold);
+        line-height: 36px;
+      }
     }
 
     .card-list-data {
       @include Flex(row, center, center, null);
-      // width: 1200px;
       width: 100%;
       position: relative;
       transition: 1s;
 
-      .arrow {
-        position: absolute;
-        cursor: pointer;
-        @include Flex(row, center, center, null);
-        border: 1px solid $gray-arrow-border;
-        border-radius: 100%;
-        background-color: $color--white;
-        width: 40px;
-        height: 40px;
-        z-index: 2;
-        &.right {
-          right: 5px;
-        }
-        &.left {
-          left: 5px;
-        }
-        &.disabled {
-          visibility: hidden;
-        }
-      }
       .card-list-wrapper {
+        position: relative;
         padding: 0 24px;
         @include Flex(row, flex-start, flex-start, 20px);
-        width: 100%;
         overflow: hidden;
-        display: flex;
         scroll-behavior: smooth;
+        @include mobile {
+          gap: 12px;
+        }
+        .arrow {
+          position: absolute;
+          cursor: pointer;
+          @include Flex(row, center, center, null);
+          border: 1px solid $gray-arrow-border;
+          border-radius: 100%;
+          background-color: $color--white;
+          width: 40px;
+          height: 40px;
+          top: 40%;
+          &.right {
+            right: 5px;
+          }
+          &.left {
+            left: 5px;
+          }
+          &.disabled {
+            visibility: hidden;
+          }
+          @include tablet {
+            visibility: hidden;
+          }
+        }
       }
     }
   }
+  .link-button-layer {
+    padding: 0 24px;
+    width: 100%;
+    @include Flex(row, center, center, null);
 
-  .post-link-button {
-    margin-top: 16px;
-    width: 280px;
-    color: $color--white;
-    background-color: $purple-600;
+    @include mobile {
+      padding: 0 20px;
+    }
+    .post-link-button {
+      margin-top: 16px;
+      width: 280px;
+      color: $color--white;
+      background-color: $purple-600;
+      @include tablet {
+        margin-top: 156px;
+        width: 100%;
+      }
+    }
   }
 }
