+++ src/utils/UseDragScroll.jsx
@@ -1,0 +1,51 @@
+import { useEffect } from 'react';
+
+function UseDragScroll(ref) {
+  useEffect(() => {
+    const element = ref.current;
+    if (!element) return;
+
+    let isDown = false;
+    let startX;
+    let scrollLeft;
+
+    const handleMouseDown = (e) => {
+      isDown = true;
+      element.classList.add('active');
+      startX = e.pageX - element.offsetLeft;
+      scrollLeft = element.scrollLeft;
+    };
+
+    const handleMouseLeave = () => {
+      isDown = false;
+      element.classList.remove('active');
+    };
+
+    const handleMouseUp = () => {
+      isDown = false;
+      element.classList.remove('active');
+    };
+
+    const handleMouseMove = (e) => {
+      if (!isDown) return;
+      e.preventDefault();
+      const x = e.pageX - element.offsetLeft;
+      const walk = (x - startX) * 2; // 스크롤 속도를 조정할 수 있습니다.
+      element.scrollLeft = scrollLeft - walk;
+    };
+
+    element.addEventListener('mousedown', handleMouseDown);
+    element.addEventListener('mouseleave', handleMouseLeave);
+    element.addEventListener('mouseup', handleMouseUp);
+    element.addEventListener('mousemove', handleMouseMove);
+
+    return () => {
+      element.removeEventListener('mousedown', handleMouseDown);
+      element.removeEventListener('mouseleave', handleMouseLeave);
+      element.removeEventListener('mouseup', handleMouseUp);
+      element.removeEventListener('mousemove', handleMouseMove);
+    };
+  }, [ref]);
+}
+
+export default UseDragScroll;
