+++ src/utils/UseWindowWidth.jsx
@@ -1,0 +1,20 @@
+import { useEffect, useState } from "react";
+
+function UseWindowWidth() {
+    const [width, setWidth] = useState(window.innerWidth);
+  
+    useEffect(() => {
+      const handleResize = () => {
+        setWidth(window.innerWidth);
+      };
+  
+      window.addEventListener('resize', handleResize);
+  
+      return () => {
+        window.removeEventListener('resize', handleResize);
+      };
+    }, []);
+    return width;
+  }
+
+  export default UseWindowWidth
